Move search::attribute::ISearchContext to searchcommon in preparation of extending IAttributeVector API.
