+++ searchcommon/src/vespa/searchcommon/common/range.h
