+++ searchlib/src/vespa/searchlib/attribute/attributevector.h
@@ -5,16 +5,16 @@
 #include "address_space_usage.h"
 #include "iattributesavetarget.h"
 #include "changevector.h"
-#include "isearchcontext.h"
 #include <vespa/fastlib/text/normwordfolder.h>
 #include <vespa/searchcommon/attribute/config.h>
+#include <vespa/searchcommon/attribute/i_search_context.h>
 #include <vespa/searchcommon/attribute/iattributevector.h>
 #include <vespa/searchcommon/attribute/status.h>
+#include <vespa/searchcommon/common/range.h>
 #include <vespa/searchcommon/common/undefinedvalues.h>
 #include <vespa/searchlib/common/address_space.h>
 #include <vespa/searchlib/common/i_compactable_lid_space.h>
 #include <vespa/searchlib/common/identifiable.h>
-#include <vespa/searchlib/common/range.h>
 #include <vespa/searchlib/common/rcuvector.h>
 #include <vespa/searchlib/queryeval/searchiterator.h>
 #include <vespa/vespalib/objects/identifiable.h>
